Draw using glDrawRangeElements.  Still using deprecated generic vertex attributes though. Fix GL texture memory leaks. Fix compiler warnings.
git-svn-id: https://www.williamfeely.info/svn/dxgl@79 8a90861a-4eca-46d5-b744-240ff16d0c4d

+++ ddraw/shadergen.h
@@ -20,7 +20,7 @@
 
 
 void SetShader(__int64 id, bool builtin);
-__int64 GetProgram();
+GLuint GetProgram();
 
 
 #endif
